+++ crates/matrix-sdk-sqlite/src/crypto_store.rs
@@ -356,6 +356,17 @@ async fn run_migrations(conn: &SqliteAsyncConn, version: u8) -> Result<()> {
         .await?;
     }
 
+    if version < 16 {
+        debug!("Upgrading database to version 16");
+        conn.with_transaction(|txn| {
+            txn.execute_batch(include_str!(
+                "../migrations/crypto_store/016_remove_old_generation_counter.sql"
+            ))?;
+            txn.set_db_version(16)
+        })
+        .await?;
+    }
+
     Ok(())
 }